Add variables to StoxBioticData from BioticData

Add variables to StoxBioticData from BioticData

AddToStoxBiotic(
  StoxBioticData,
  BioticData,
  VariableNames = character(),
  SplitTableAllocation = c("Default", "Lowest", "Highest")
)

Arguments

StoxBioticData

StoxBioticData.

BioticData

BioticData.

VariableNames

A character vector with names of the variables to add from the BioticData.

SplitTableAllocation

A string indicating how to split tables of the BioticData into tables of the StoxBiotic format. See Details.

Value

An object of StoX data type StoxBioticData.

Details

The default SplitTableAllocation allocates variables to the StoxBiotic tables according to the mapping defined in

RstoxData::stoxBioticObject$complexMaps.

When SplitTableAllocation is "Lowest" or "Highest" variables are allocated to the lowest or highest table, respectively, as defined in

RstoxData::stoxBioticObject$complexMaps_lowestTable

and

RstoxData::stoxBioticObject$complexMaps_highestTable,

respectively. In NMDBiotic the tables that are split are fishstation -> Station/Haul (where the first table is the highest and the second is the lowest in the General sampling hierarchy) and catchsample -> SpeciesCategory/Sample. In ICESBiotic the tables that are split are Haul -> Station/Haul and Catch -> SpeciesCategory/Sample for NMDBiotic.