Convert StoxBiotic variables

This function converts specific vaiables of specific tables of StoxBioticData to used defined new values.

ConvertStoxBioticVariables(
  StoxBioticData,
  ConversionMethod = c("Table", "PreDefined"),
  VariableConversionTable = data.table::data.table(),
  StoxBioticVariableConversion
)

Arguments

StoxBioticData
Input StoxBioticData data.

ConversionMethod
Character: A string naming the method to use, one of "Table", for providing the old and new values of the variables in the table VariableConversionTable; and "PreDefined" for providing the table using the argument BioticVariableConversion.

VariableConversionTable
A table of the columns TableName, VariableName, Value and NewValue, specifying the conversion from Value to NewValue for any number of variables of any number of tables of BioticData.

StoxBioticVariableConversion
The (optional) StoxBioticVariableConversion process data.

Value

A StoxBioticData object.