Copy a variable of ICESDatrasData

This function copies a variable to another (possibly existing) variable of ICESDatrasData .

CopyICESDatras(
  ICESDatrasData,
  FromVariable = character(),
  ToVariable = character(),
  Overwrite = FALSE,
  PreserveClass = TRUE
)

Arguments

ICESDatrasData

ICESDatrasData.

FromVariable

The name of the variable to copy.

ToVariable

The name of the (possibly existing) variable insert the values in FromVariable to.

Overwrite

Logical: If TRUE overwrite the existing variable, if ToVariable is the name of an existing variable in the same table as FromVariable.

PreserveClass

Logical: Only used if Overwrite is TRUE. If PreserveClass is TRUE (the default) do not convert the class of the variable given by ToVariable to the class of the variable given by FromVariable.

Value

A ICESDatrasData object.