StoxAcoustic data format.

The StoxAcoustic data format is defined by StoX as a common format to which data from different acoustic sampling formats are converted, guaranteeing consistent interpretation and documentation of all its variables.

Details

The StoxAcoustic format is defined according to the general sampling hierarchy of StoX which is used as a basis for both the StoxcBiotic and StoxAcoustic format. The variables of the StoxAcoustic format are given by the tables below, where the columns NMDEchosounder and ICESAcoustic show the origin of variable in AcousticData read from NMDEchosounder https://www.imr.no/formats/nmdechosounder/ and ICESAcoustic https://www.ices.dk/data/data-portals/Pages/acoustic.aspx files, respectively (see AcousticData). Variables with non-unique names across tables of the BioticData are written as variableName.tableName, and "-" and "/" denotes concatenation for character type variables with "-" and "/" as separation character, respectively. Units are given with shortname of the StoxUnits:

Cruise level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

CruiseKey	Key of the Cruise table	None	Character	"2021105"	cruise	LocalID

Cruise	Unique Cruise identifier	None	Character	"2021105"	Same as CruiseKey. Can be translated.	Same as CruiseKey. Can be translated.

Platform	Data collection platform identifier	None	Character	"1019"	platform	Platform

Log level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

LogKey	Key of the Log level, given as ISO 8601 date time string	ISO8601	Character	"2020-04-24T03:51:26.000Z"	ISO 8601 formatted start_time	ISO 8601 formatted Time

Log	Sailed distance	nmi	Numeric	123.1	log_start	Distance

EDSU	Unique elementary distance sampling unit (EDSU/Log) identifier	None	Character	"2020821/2020-04-24T03:51:26.000Z"	CruiseKey-LogKey	CruiseKey-LogKey

DateTime	UTC time at start of the EDSU, stored as POSIXct	ISO8601	Character	2020-09-09T01:02:03.456Z	start_time	Time

Longitude	Longitude at LogOrigin	degree east	Numeric	62.5	lon_start	Longitude

Latitude	Latitude at LogOrigin	degree north	Numeric	5.1	lat_start	Latitude

LogOrigin	Origin of the Longitude/Latitude, one of "start", "middle", "end"	None	Character	"start"	"start"	LogOrigin

Longitude2	Longitude at LogOrigin2	degree east	Numeric	62.6	lon_stop	Longitude2

Latitude2	Latitude at LogOrigin2	degree north	Numeric	5.2	lat_stop	Latitude2

LogOrigin2	Origin of the Longitude/Latitude, one of "start", "middle", "end"	None	Character	"end"	"end"	LogOrigin2

LogDistance	The length of the EDSU	nmi	Numeric	0.1	integrator_dist	PingAxisInterval

LogDuration	The duration of the EDSU	s	Numeric	36	stop_time - start_time	NA

EffectiveLogDistance	The effective length of the EDSU	nmi	Numeric	0.09	Same as LogDistance.	Same as LogDistance

BottomDepth	BottomDepth at the EDSU	m	Numeric	123.2	NA	BottomDepth

Beam level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

BeamKey	Key of the Beam level	None	Character	"38000/2"	freq/transceiver	ID.Instrument

Beam	Unique Beam identifier	None	Character	"38000/2"	Same as BeamKey. Can be translated.	Same as BeamKey. Can be translated.

Frequency	The acoustic frequency of the Beam	hertz	Numeric	38000	freq	Frequency

AcousticCategory level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

AcousticCategoryKey	Key of the AcousticCategory level	None	Character	"HER"	acocat	SaCategory

AcousticCategory	The acousic category	None	Character	"Herring"	Same as AcousticCategoryKey Can be translated.	Same as AcousticCategoryKey Can be translated.

ChannelReference level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

ChannelReferenceKey	Key of the ChannelReference level	None	Character	"P"	type	"P"

ChannelReferenceType	Unique ChannelReference identifier	None	Character	"P"	Same as ChannelReferenceKey	Same as ChannelReferenceKey

ChannelReferenceDepth	The depth of the ChannelReference origin. 0 for pelagic channels. Not yet given for bottom channels, as BottomDepth is not yet defined for NMDEchosounder data	m	Numeric	0	0 if ChannelReferenceType == "P", NA if ChannelReferenceType == "B" (see BottomDepth)	0

ChannelReferenceTilt	The tilt angle of the beam, where 180 is vertically downwards and 0 is vertically upwards	degree	Numeric	180	180 if ChannelReferenceType == "P", 0 if ChannelReferenceType == "B"	Interpreted from TransducerOrientation

NASC level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

NASCKey	Key of the NASC level	None	Character	"2"	ch	ChannelDepthUpper/ChannelDepthLower

Channel	Unique NASC depth channel identifier	None	Character	"2"	Same as NASCKey Can be translated.	Same as NASCKey Can be translated.

MaxChannelRange	The maximum range of the channel. Translates to the lower channel depth for vertically downwards oriented echosounder.	m	Numeric	40	pel_ch_thickness * ch	MaxChannelRange

MinChannelRange	The minimum range of the channel. Translates to the upper channel depth for vertically downwards oriented echosounder.	m	Numeric	30	pel_ch_thickness * (ch - 1)	MinChannelRange

NASC	The nautical area scattering coefficient.	m^2/nmi^2	Numeric	59.24813	sa	Value