StoxBiotic data format.

The StoxBiotic data format is defined by StoX as a common format to which data from different biotic sampling formats are converted, guaranteeing consistent interpretation and documentation of all its variables.

Details

The StoxBiotic format is defined according to the general sampling hierarchy of StoX which is used as a basis for both the StoxcBiotic and StoxAcoustic format. The variables of the StoxBiotic format are given by the tables below:

Cruise level:

Variable	Description	Unit	Data type	Example

CruiseKey	Key of the Cruise table	None	Character	2021105

Cruise	Unique Cruise identifier	None	Character type	2021105

Platform	Data collection platform identifier	None	Character	1019

Station level:

Variable	Description	Unit	Data type	Example

StationKey	Key of the Station level	None	1	Station

Unique Station identifier	None	2021105-1	CatchPlatform	Platform performing the actual sampling (can be different from the data collection platform)

None	Character	1019	DateTime	UTC time at start of the station

ISO 8601 format with milliseconds	Character	2020-09-09T01:02:03.456Z	Longitude	Longitude at start of the station

Decimal degrees	Numeric	62.5	Latitude	Latitude at start of the station

Decimal degrees	Numeric	5.1	BottomDepth	Bottom depth at start of the station

Haul level:

Variable	Description	Unit	Data type	Example

HaulKey	Key of the Haul level	None	Character	2

Haul	Unique Haul identifier	None	Character	2021105-1-2

Gear	Identifier of the gear	None	Character	3270

TowDistance	Distance between start and end of the haul	Nautical miles	Numeric	1.5

EffectiveTowDistance	Effective tow distance	Nautical miles	Numeric	1.5

MinHaulDepth	Minimum depth of the haul (trawl headline)	m	Numeric	65

MaxHaulDepth	Maximum depth of the haul (trawl headline)	m	Numeric	35

VerticalNetOpening	Vertical span of the net	m	Numeric	23

HorizontalNetOpening	Vertical span of the net	m	Numeric	105

TrawlDoorSpread	Distance between the trawl doors	m	Numeric	125

SpeciesCategory level:

Variable	Description	Unit	Data type	Example

SpeciesCategoryKey	Key of the SpeciesCategory level	None	Character	126417

SpeciesCategory	The species category	None	Character	Herring

Sample level:

Variable	Description	Unit	Data type	Example

SampleKey	Key of the Sample level	None	Character

Sample	Unique Sample identifier	None	Character

CatchFractionWeight	Total weight of the catch SpeciesCategory and sub category (fractions such as juveniles and adults)	Kg	Numeric	49.9

CatchFractionCount	Total number of individuals of the catch SpeciesCategory and sub category (fractions such as juveniles and adults)	None	Numeric	295

SampleWeight	Total weight of the sample for individual measurements	Kg	Numeric	4.6

SampleCount	Size of the sample for individual measurements	None	Numeric	100

Individual level:

Variable	Description	Unit	Data type	Example

IndividualKey	Key of the Individual level	None	Character

Individual	Unique Individual identifier	None	Character

IndividualRoundWeight	Round weight (the whole fish) fo the individual	g	Numeric	123

IndividualTotalLength	Total length (from snoute to end of fin)	cm	Numeric	14.5

LengthResolution	Resolution of IndividualTotalLength	cm	Numeric	0.5

WeightMeasurement		None	Character

IndividualAge	Age of an individual	year	Numeric	3

IndividualSex	sex of an individual	F is female, M is male	Character	F