Mapping from BioticData to StoxBiotic data for different input file formats

The StoxBioticFormat contains a fixed mapping between variables from the different input biotic data formats to the variables of the StoxBioticData. For variables that are not supported by StoxBiotic the function AddToStoxBiotic can be used to add these to the StoxBioticData. The function RedefineStoxBiotic can also be used to redefine variables in a StoxBioticData object. Both of these functions add data to the StoxBioticData object using the described in the Details.

Details

The mapping from the different biotic data format to StoxBiotic is different for different levels Some levels are directly mapped to a levels in StoxBioticData, whereas other levels are split into two levels of the StoxBioticData.

The following table shows levels of different formats that are mapped directly, i.e., all variables of the source levels are mapped to the desination levels:

Format	Level	DesinationLevel	icesBiotic	Cruise	Cruise

icesBiotic	Biology	Individual	nmdbioticv1.1	mission	Cruise

nmdbioticv1.1	individual	Individual	nmdbioticv1.4	mission	Cruise

nmdbioticv1.4	individual	Individual	nmdbioticv3	mission	Cruise

nmdbioticv3	individual	Individual	nmdbioticv3.1	mission	Cruise

Note that the age from the agedetermination level of nmdbioticv1.1, nmdbioticv1.4, nmdbioticv3 and nmdbioticv3.1 is selected based on the preferredagereading in the individual level and placed in the Individual level of StoxBioticData.

The mapping from one level to two levels is shown in the table below. The parameter SplitTableAllocation in AddToStoxBiotic and RedefineStoxBiotic is used to determine the mapping. The columns of the mapping are as follows:

Format

The biotic data format

Level

The table in the BioticData

Variable

The variable in the BioticData

DesinationLevel_Default

The table of the StoxBioticData to which the variable is placed by default (SplitTableAllocation = "Default")

DesinationLevel_Lowest

The table of the StoxBioticData to which the variable is placed when SplitTableAllocation = "Lowest"

DesinationLevel_Highest

The table of the StoxBioticData to which the variable is placed when SplitTableAllocation = "Highest"

Mapping from one level in BioticData to two levels in StoxBioticData

Format	Level	Variable	DesinationLevel_Default	DesinationLevel_Lowest	DesinationLevel_Highest

icesBiotic	Haul	CruiseKey	Station	Haul	Station

icesBiotic	Haul	StationKey	Station	Haul	Station

icesBiotic	Haul	Gear	Haul	Haul	Station

icesBiotic	Haul	Number	Station	Haul	Station

icesBiotic	Haul	StationName	Station	Haul	Station

icesBiotic	Haul	StartTime	Station	Haul	Station

icesBiotic	Haul	Duration	Haul	Haul	Station

icesBiotic	Haul	Validity	Haul	Haul	Station

icesBiotic	Haul	Platform	Station	Haul	Station

icesBiotic	Haul	StartLatitude	Station	Haul	Station

icesBiotic	Haul	StartLongitude	Station	Haul	Station

icesBiotic	Haul	StopLatitude	Station	Haul	Station

icesBiotic	Haul	StopLongitude	Station	Haul	Station

icesBiotic	Haul	StatisticalRectangle	Station	Haul	Station

icesBiotic	Haul	MinTrawlDepth	Haul	Haul	Station

icesBiotic	Haul	MaxTrawlDepth	Haul	Haul	Station

icesBiotic	Haul	BottomDepth	Station	Haul	Station

icesBiotic	Haul	CruiseKey	Haul	Haul	Station

icesBiotic	Haul	StationKey	Haul	Haul	Station

icesBiotic	Haul	HaulKey	Haul	Haul	Station

icesBiotic	Haul	Number	Haul	Haul	Station

icesBiotic	Haul	Gear	Haul	Haul	Station

icesBiotic	Haul	Distance	Haul	Haul	Station

icesBiotic	Haul	Netopening	Haul	Haul	Station

icesBiotic	Haul	CodendMesh	Haul	Haul	Station

icesBiotic	Haul	SweepLength	Haul	Haul	Station

icesBiotic	Haul	GearExceptions	Haul	Haul	Station

icesBiotic	Haul	DoorType	Haul	Haul	Station

icesBiotic	Haul	WarpLength	Haul	Haul	Station

icesBiotic	Haul	WarpDiameter	Haul	Haul	Station

icesBiotic	Haul	WarpDensity	Haul	Haul	Station

icesBiotic	Haul	DoorSurface	Haul	Haul	Station

icesBiotic	Haul	DoorWeight	Haul	Haul	Station

icesBiotic	Haul	DoorSpread	Haul	Haul	Station

icesBiotic	Haul	WingSpread	Haul	Haul	Station

icesBiotic	Haul	Buoyancy	Haul	Haul	Station

icesBiotic	Haul	KiteArea	Haul	Haul	Station

icesBiotic	Haul	GroundRopeWeight	Haul	Haul	Station

icesBiotic	Haul	Rigging	Haul	Haul	Station

icesBiotic	Haul	Tickler	Haul	Haul	Station

icesBiotic	Haul	HydrographicStationID	Station	Haul	Station

icesBiotic	Haul	TowDirection	Station	Haul	Station

icesBiotic	Haul	SpeedGround	Station	Haul	Station

icesBiotic	Haul	SpeedWater	Station	Haul	Station

icesBiotic	Haul	WindDirection	Station	Haul	Station

icesBiotic	Haul	WindSpeed	Station	Haul	Station

icesBiotic	Haul	SwellDirection	Station	Haul	Station

icesBiotic	Haul	SwellHeight	Station	Haul	Station

icesBiotic	Catch	DataType	Sample	Sample	SpeciesCategory

icesBiotic	Catch	CruiseKey	SpeciesCategory	Sample	SpeciesCategory

icesBiotic	Catch	StationKey	SpeciesCategory	Sample	SpeciesCategory

icesBiotic	Catch	HaulKey	SpeciesCategory	Sample	SpeciesCategory

icesBiotic	Catch	SpeciesCategoryKey	SpeciesCategory	Sample	SpeciesCategory

icesBiotic	Catch	SpeciesCode	SpeciesCategory	Sample	SpeciesCategory

icesBiotic	Catch	SpeciesValidity	SpeciesCategory	Sample	SpeciesCategory

icesBiotic	Catch	SpeciesCategory	SpeciesCategory	Sample	SpeciesCategory

icesBiotic	Catch	SpeciesCategoryNumber	SpeciesCategory	Sample	SpeciesCategory

icesBiotic	Catch	WeightUnit	Sample	Sample	SpeciesCategory

icesBiotic	Catch	SpeciesCategoryWeight	SpeciesCategory	Sample	SpeciesCategory

icesBiotic	Catch	SpeciesSex	SpeciesCategory	Sample	SpeciesCategory

icesBiotic	Catch	CruiseKey	Sample	Sample	SpeciesCategory

icesBiotic	Catch	StationKey	Sample	Sample	SpeciesCategory

icesBiotic	Catch	HaulKey	Sample	Sample	SpeciesCategory

icesBiotic	Catch	SpeciesCategoryKey	Sample	Sample	SpeciesCategory

icesBiotic	Catch	SampleKey	Sample	Sample	SpeciesCategory

icesBiotic	Catch	SpeciesCategoryWeight	Sample	Sample	SpeciesCategory

icesBiotic	Catch	SpeciesCategoryNumber	Sample	Sample	SpeciesCategory

icesBiotic	Catch	SubsampledNumber	Sample	Sample	SpeciesCategory

icesBiotic	Catch	SubsamplingFactor	Sample	Sample	SpeciesCategory

icesBiotic	Catch	SubsampleWeight	Sample	Sample	SpeciesCategory

icesBiotic	Catch	LengthCode	Sample	Sample	SpeciesCategory

icesBiotic	Catch	LengthClass	Sample	Sample	SpeciesCategory

icesBiotic	Catch	LengthType	Sample	Sample	SpeciesCategory

icesBiotic	Catch	NumberAtLength	Sample	Sample	SpeciesCategory

icesBiotic	Catch	WeightAtLength	Sample	Sample	SpeciesCategory

nmdbioticv1.1	fishstation	serialno	Station	Haul	Station

nmdbioticv1.1	fishstation	nation	Station	Haul	Station

nmdbioticv1.1	fishstation	platform.fishstation	Station	Haul	Station

nmdbioticv1.1	fishstation	station	Station	Haul	Station

nmdbioticv1.1	fishstation	fixedstation	Station	Haul	Station

nmdbioticv1.1	fishstation	startdate.fishstation	Station	Haul	Station

nmdbioticv1.1	fishstation	starttime	Station	Haul	Station

nmdbioticv1.1	fishstation	stopdate.fishstation	Station	Haul	Station

nmdbioticv1.1	fishstation	stoptime	Station	Haul	Station

nmdbioticv1.1	fishstation	stationtype	Station	Haul	Station

nmdbioticv1.1	fishstation	latitudestart	Station	Haul	Station

nmdbioticv1.1	fishstation	longitudestart	Station	Haul	Station

nmdbioticv1.1	fishstation	latitudeend	Station	Haul	Station

nmdbioticv1.1	fishstation	longitudeend	Station	Haul	Station

nmdbioticv1.1	fishstation	system	Station	Haul	Station

nmdbioticv1.1	fishstation	area	Station	Haul	Station

nmdbioticv1.1	fishstation	location	Station	Haul	Station

nmdbioticv1.1	fishstation	bottomdepthstart	Station	Haul	Station

nmdbioticv1.1	fishstation	bottomdepthstop	Station	Haul	Station

nmdbioticv1.1	fishstation	fishingdepthmax	Haul	Haul	Station

nmdbioticv1.1	fishstation	fishingdepthmin	Haul	Haul	Station

nmdbioticv1.1	fishstation	fishingdepthmean	Haul	Haul	Station

nmdbioticv1.1	fishstation	gearno	Haul	Haul	Station

nmdbioticv1.1	fishstation	gear	Haul	Haul	Station

nmdbioticv1.1	fishstation	gearcount	Haul	Haul	Station

nmdbioticv1.1	fishstation	direction	Haul	Haul	Station

nmdbioticv1.1	fishstation	gearspeed	Haul	Haul	Station

nmdbioticv1.1	fishstation	startlog	Station	Haul	Station

nmdbioticv1.1	fishstation	stoplog	Station	Haul	Station

nmdbioticv1.1	fishstation	distance	Haul	Haul	Station

nmdbioticv1.1	fishstation	gearcondition	Haul	Haul	Station

nmdbioticv1.1	fishstation	trawlquality	Haul	Haul	Station

nmdbioticv1.1	fishstation	trawlopening	Haul	Haul	Station

nmdbioticv1.1	fishstation	trawlopeningsd	Haul	Haul	Station

nmdbioticv1.1	fishstation	doorspread	Haul	Haul	Station

nmdbioticv1.1	fishstation	doorspreadsd	Haul	Haul	Station

nmdbioticv1.1	fishstation	wirelength	Haul	Haul	Station

nmdbioticv1.1	fishstation	soaktime	Haul	Haul	Station

nmdbioticv1.1	fishstation	tripno	Station	Haul	Station

nmdbioticv1.1	fishstation	fishabundance	Station	Haul	Station

nmdbioticv1.1	fishstation	fishdistribution	Station	Haul	Station

nmdbioticv1.1	fishstation	landingsite	Station	Haul	Station

nmdbioticv1.1	fishstation	fishingground	Station	Haul	Station

nmdbioticv1.1	fishstation	countofvessels	Station	Haul	Station

nmdbioticv1.1	fishstation	dataquality	Haul	Haul	Station

nmdbioticv1.1	fishstation	haulvalidity	Haul	Haul	Station

nmdbioticv1.1	fishstation	flora	Station	Haul	Station

nmdbioticv1.1	fishstation	vegetationcover	Station	Haul	Station

nmdbioticv1.1	fishstation	visibility	Station	Haul	Station

nmdbioticv1.1	fishstation	flowcount	Station	Haul	Station

nmdbioticv1.1	fishstation	flowconst	Station	Haul	Station

nmdbioticv1.1	fishstation	comment	Station	Haul	Station

nmdbioticv1.1	catchsample	species	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	samplenumber	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	noname	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	aphia	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	sampletype	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	group	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	conservation	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	producttype	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	weight	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	volume	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	count	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	abundancecategory	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	sampleproducttype	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	lengthmeasurement	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	lengthsampleweight	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	lengthsamplevolume	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	lengthsamplecount	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	specimensamplecount	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	raisingfactor	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	agingstructure	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	parasite	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	stomach	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	genetics	Sample	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	foreignobject	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv1.1	catchsample	comment.catchsample	Sample	Sample	SpeciesCategory

nmdbioticv1.4	fishstation	serialno	Station	Haul	Station

nmdbioticv1.4	fishstation	nation	Station	Haul	Station

nmdbioticv1.4	fishstation	platform.fishstation	Station	Haul	Station

nmdbioticv1.4	fishstation	station	Station	Haul	Station

nmdbioticv1.4	fishstation	fixedstation	Station	Haul	Station

nmdbioticv1.4	fishstation	startdate.fishstation	Station	Haul	Station

nmdbioticv1.4	fishstation	starttime	Station	Haul	Station

nmdbioticv1.4	fishstation	stopdate.fishstation	Station	Haul	Station

nmdbioticv1.4	fishstation	stoptime	Station	Haul	Station

nmdbioticv1.4	fishstation	stationtype	Station	Haul	Station

nmdbioticv1.4	fishstation	latitudestart	Station	Haul	Station

nmdbioticv1.4	fishstation	longitudestart	Station	Haul	Station

nmdbioticv1.4	fishstation	latitudeend	Station	Haul	Station

nmdbioticv1.4	fishstation	longitudeend	Station	Haul	Station

nmdbioticv1.4	fishstation	system	Station	Haul	Station

nmdbioticv1.4	fishstation	area	Station	Haul	Station

nmdbioticv1.4	fishstation	location	Station	Haul	Station

nmdbioticv1.4	fishstation	bottomdepthstart	Station	Haul	Station

nmdbioticv1.4	fishstation	bottomdepthstop	Station	Haul	Station

nmdbioticv1.4	fishstation	fishingdepthmax	Haul	Haul	Station

nmdbioticv1.4	fishstation	fishingdepthmin	Haul	Haul	Station

nmdbioticv1.4	fishstation	fishingdepthmean	Haul	Haul	Station

nmdbioticv1.4	fishstation	gearno	Haul	Haul	Station

nmdbioticv1.4	fishstation	gear	Haul	Haul	Station

nmdbioticv1.4	fishstation	gearcount	Haul	Haul	Station

nmdbioticv1.4	fishstation	direction	Haul	Haul	Station

nmdbioticv1.4	fishstation	gearspeed	Haul	Haul	Station

nmdbioticv1.4	fishstation	startlog	Station	Haul	Station

nmdbioticv1.4	fishstation	stoplog	Station	Haul	Station

nmdbioticv1.4	fishstation	distance	Haul	Haul	Station

nmdbioticv1.4	fishstation	gearcondition	Haul	Haul	Station

nmdbioticv1.4	fishstation	trawlquality	Haul	Haul	Station

nmdbioticv1.4	fishstation	trawlopening	Haul	Haul	Station

nmdbioticv1.4	fishstation	trawlopeningsd	Haul	Haul	Station

nmdbioticv1.4	fishstation	doorspread	Haul	Haul	Station

nmdbioticv1.4	fishstation	doorspreadsd	Haul	Haul	Station

nmdbioticv1.4	fishstation	wirelength	Haul	Haul	Station

nmdbioticv1.4	fishstation	soaktime	Haul	Haul	Station

nmdbioticv1.4	fishstation	tripno	Station	Haul	Station

nmdbioticv1.4	fishstation	fishabundance	Station	Haul	Station

nmdbioticv1.4	fishstation	fishdistribution	Station	Haul	Station

nmdbioticv1.4	fishstation	landingsite	Station	Haul	Station

nmdbioticv1.4	fishstation	fishingground	Station	Haul	Station

nmdbioticv1.4	fishstation	countofvessels	Station	Haul	Station

nmdbioticv1.4	fishstation	dataquality	Haul	Haul	Station

nmdbioticv1.4	fishstation	haulvalidity	Haul	Haul	Station

nmdbioticv1.4	fishstation	flora	Station	Haul	Station

nmdbioticv1.4	fishstation	vegetationcover	Station	Haul	Station

nmdbioticv1.4	fishstation	visibility	Station	Haul	Station

nmdbioticv1.4	fishstation	flowcount	Station	Haul	Station

nmdbioticv1.4	fishstation	flowconst	Station	Haul	Station

nmdbioticv1.4	fishstation	waterlevel	Station	Haul	Station

nmdbioticv1.4	fishstation	winddirection	Station	Haul	Station

nmdbioticv1.4	fishstation	windspeed	Station	Haul	Station

nmdbioticv1.4	fishstation	clouds	Station	Haul	Station

nmdbioticv1.4	fishstation	sea	Station	Haul	Station

nmdbioticv1.4	fishstation	weather	Station	Haul	Station

nmdbioticv1.4	fishstation	comment	Station	Haul	Station

nmdbioticv1.4	catchsample	species	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	samplenumber	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	noname	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	aphia	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	sampletype	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	group	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	conservation	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	producttype	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	weight	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	volume	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	count	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	abundancecategory	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	sampleproducttype	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	lengthmeasurement	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	lengthsampleweight	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	lengthsamplevolume	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	lengthsamplecount	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	specimensamplecount	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	raisingfactor	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	agingstructure	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	parasite	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	stomach	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	genetics	Sample	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	foreignobject	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv1.4	catchsample	comment.catchsample	Sample	Sample	SpeciesCategory

nmdbioticv3.1	fishstation	serialnumber	Station	Haul	Station

nmdbioticv3.1	fishstation	nation	Station	Haul	Station

nmdbioticv3.1	fishstation	catchplatform	Station	Haul	Station

nmdbioticv3.1	fishstation	station	Station	Haul	Station

nmdbioticv3.1	fishstation	fixedstation	Station	Haul	Station

nmdbioticv3.1	fishstation	stationstartdate	Station	Haul	Station

nmdbioticv3.1	fishstation	stationstarttime	Station	Haul	Station

nmdbioticv3.1	fishstation	stationstopdate	Station	Haul	Station

nmdbioticv3.1	fishstation	stationstoptime	Station	Haul	Station

nmdbioticv3.1	fishstation	stationtype	Station	Haul	Station

nmdbioticv3.1	fishstation	latitudestart	Station	Haul	Station

nmdbioticv3.1	fishstation	longitudestart	Station	Haul	Station

nmdbioticv3.1	fishstation	latitudeend	Station	Haul	Station

nmdbioticv3.1	fishstation	longitudeend	Station	Haul	Station

nmdbioticv3.1	fishstation	system	Station	Haul	Station

nmdbioticv3.1	fishstation	area	Station	Haul	Station

nmdbioticv3.1	fishstation	location	Station	Haul	Station

nmdbioticv3.1	fishstation	bottomdepthstart	Station	Haul	Station

nmdbioticv3.1	fishstation	bottomdepthstop	Station	Haul	Station

nmdbioticv3.1	fishstation	bottomdepthmean	Station	Haul	Station

nmdbioticv3.1	fishstation	fishingdepthstart	Haul	Haul	Station

nmdbioticv3.1	fishstation	fishingdepthstop	Haul	Haul	Station

nmdbioticv3.1	fishstation	fishingdepthcount	Haul	Haul	Station

nmdbioticv3.1	fishstation	fishingdepthmax	Haul	Haul	Station

nmdbioticv3.1	fishstation	fishingdepthmin	Haul	Haul	Station

nmdbioticv3.1	fishstation	fishingdepthmean	Haul	Haul	Station

nmdbioticv3.1	fishstation	fishingdepthtemperature	Haul	Haul	Station

nmdbioticv3.1	fishstation	fishingbait	Haul	Haul	Station

nmdbioticv3.1	fishstation	gearno	Haul	Haul	Station

nmdbioticv3.1	fishstation	sweeplength	Haul	Haul	Station

nmdbioticv3.1	fishstation	gear	Haul	Haul	Station

nmdbioticv3.1	fishstation	gearcount	Haul	Haul	Station

nmdbioticv3.1	fishstation	direction	Haul	Haul	Station

nmdbioticv3.1	fishstation	gearflow	Haul	Haul	Station

nmdbioticv3.1	fishstation	vesselspeed	Haul	Haul	Station

nmdbioticv3.1	fishstation	logstart	Station	Haul	Station

nmdbioticv3.1	fishstation	logstop	Station	Haul	Station

nmdbioticv3.1	fishstation	distance	Haul	Haul	Station

nmdbioticv3.1	fishstation	gearcondition	Haul	Haul	Station

nmdbioticv3.1	fishstation	samplequality	Haul	Haul	Station

nmdbioticv3.1	fishstation	verticaltrawlopening	Haul	Haul	Station

nmdbioticv3.1	fishstation	verticaltrawlopeningsd	Haul	Haul	Station

nmdbioticv3.1	fishstation	trawldoortype	Haul	Haul	Station

nmdbioticv3.1	fishstation	trawldoorarea	Haul	Haul	Station

nmdbioticv3.1	fishstation	trawldoorweight	Haul	Haul	Station

nmdbioticv3.1	fishstation	trawldoorspread	Haul	Haul	Station

nmdbioticv3.1	fishstation	trawldoorspreadsd	Haul	Haul	Station

nmdbioticv3.1	fishstation	wingspread	Haul	Haul	Station

nmdbioticv3.1	fishstation	wingspreadsd	Haul	Haul	Station

nmdbioticv3.1	fishstation	wirelength	Haul	Haul	Station

nmdbioticv3.1	fishstation	wirediameter	Haul	Haul	Station

nmdbioticv3.1	fishstation	wiredensity	Haul	Haul	Station

nmdbioticv3.1	fishstation	soaktime	Haul	Haul	Station

nmdbioticv3.1	fishstation	tripno	Station	Haul	Station

nmdbioticv3.1	fishstation	fishabundance	Station	Haul	Station

nmdbioticv3.1	fishstation	fishdistribution	Station	Haul	Station

nmdbioticv3.1	fishstation	landingsite	Station	Haul	Station

nmdbioticv3.1	fishstation	fishingground	Station	Haul	Station

nmdbioticv3.1	fishstation	vesselcount	Station	Haul	Station

nmdbioticv3.1	fishstation	dataquality	Haul	Haul	Station

nmdbioticv3.1	fishstation	haulvalidity	Haul	Haul	Station

nmdbioticv3.1	fishstation	flora	Station	Haul	Station

nmdbioticv3.1	fishstation	vegetationcover	Station	Haul	Station

nmdbioticv3.1	fishstation	visibility	Station	Haul	Station

nmdbioticv3.1	fishstation	waterlevel	Station	Haul	Station

nmdbioticv3.1	fishstation	winddirection	Station	Haul	Station

nmdbioticv3.1	fishstation	windspeed	Station	Haul	Station

nmdbioticv3.1	fishstation	clouds	Station	Haul	Station

nmdbioticv3.1	fishstation	sea	Station	Haul	Station

nmdbioticv3.1	fishstation	weather	Station	Haul	Station

nmdbioticv3.1	fishstation	stationcomment	Station	Haul	Station

nmdbioticv3.1	catchsample	commonname	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	catchcategory	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	catchpartnumber	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	aphia	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	scientificname	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	identification	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	foreignobject	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	sampletype	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	group	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	conservation	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	catchproducttype	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	raisingfactor	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	catchweight	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	catchvolume	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	catchcount	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	abundancecategory	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	sampleproducttype	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	lengthmeasurement	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	lengthsampleweight	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	lengthsamplevolume	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	lengthsamplecount	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	specimensamplecount	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	agesamplecount	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	agingstructure	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	parasite	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	stomach	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	intestine	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	tissuesample	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	samplerecipient	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	catchcomment	Sample	Sample	SpeciesCategory

nmdbioticv3.1	catchsample	catchsampleid	Sample	Sample	SpeciesCategory

nmdbioticv3	fishstation	serialnumber	Station	Haul	Station

nmdbioticv3	fishstation	nation	Station	Haul	Station

nmdbioticv3	fishstation	catchplatform	Station	Haul	Station

nmdbioticv3	fishstation	station	Station	Haul	Station

nmdbioticv3	fishstation	fixedstation	Station	Haul	Station

nmdbioticv3	fishstation	stationstartdate	Station	Haul	Station

nmdbioticv3	fishstation	stationstarttime	Station	Haul	Station

nmdbioticv3	fishstation	stationstopdate	Station	Haul	Station

nmdbioticv3	fishstation	stationstoptime	Station	Haul	Station

nmdbioticv3	fishstation	stationtype	Station	Haul	Station

nmdbioticv3	fishstation	latitudestart	Station	Haul	Station

nmdbioticv3	fishstation	longitudestart	Station	Haul	Station

nmdbioticv3	fishstation	latitudeend	Station	Haul	Station

nmdbioticv3	fishstation	longitudeend	Station	Haul	Station

nmdbioticv3	fishstation	system	Station	Haul	Station

nmdbioticv3	fishstation	area	Station	Haul	Station

nmdbioticv3	fishstation	location	Station	Haul	Station

nmdbioticv3	fishstation	bottomdepthstart	Station	Haul	Station

nmdbioticv3	fishstation	bottomdepthstop	Station	Haul	Station

nmdbioticv3	fishstation	bottomdepthmean	Station	Haul	Station

nmdbioticv3	fishstation	fishingdepthstart	Haul	Haul	Station

nmdbioticv3	fishstation	fishingdepthstop	Haul	Haul	Station

nmdbioticv3	fishstation	fishingdepthcount	Haul	Haul	Station

nmdbioticv3	fishstation	fishingdepthmax	Haul	Haul	Station

nmdbioticv3	fishstation	fishingdepthmin	Haul	Haul	Station

nmdbioticv3	fishstation	fishingdepthmean	Haul	Haul	Station

nmdbioticv3	fishstation	fishingdepthtemperature	Haul	Haul	Station

nmdbioticv3	fishstation	gearno	Haul	Haul	Station

nmdbioticv3	fishstation	sweeplength	Haul	Haul	Station

nmdbioticv3	fishstation	gear	Haul	Haul	Station

nmdbioticv3	fishstation	gearcount	Haul	Haul	Station

nmdbioticv3	fishstation	direction	Haul	Haul	Station

nmdbioticv3	fishstation	gearflow	Haul	Haul	Station

nmdbioticv3	fishstation	vesselspeed	Haul	Haul	Station

nmdbioticv3	fishstation	logstart	Station	Haul	Station

nmdbioticv3	fishstation	logstop	Station	Haul	Station

nmdbioticv3	fishstation	distance	Haul	Haul	Station

nmdbioticv3	fishstation	gearcondition	Haul	Haul	Station

nmdbioticv3	fishstation	samplequality	Haul	Haul	Station

nmdbioticv3	fishstation	verticaltrawlopening	Haul	Haul	Station

nmdbioticv3	fishstation	verticaltrawlopeningsd	Haul	Haul	Station

nmdbioticv3	fishstation	trawldoortype	Haul	Haul	Station

nmdbioticv3	fishstation	trawldoorarea	Haul	Haul	Station

nmdbioticv3	fishstation	trawldoorweight	Haul	Haul	Station

nmdbioticv3	fishstation	trawldoorspread	Haul	Haul	Station

nmdbioticv3	fishstation	trawldoorspreadsd	Haul	Haul	Station

nmdbioticv3	fishstation	wingspread	Haul	Haul	Station

nmdbioticv3	fishstation	wingspreadsd	Haul	Haul	Station

nmdbioticv3	fishstation	wirelength	Haul	Haul	Station

nmdbioticv3	fishstation	wirediameter	Haul	Haul	Station

nmdbioticv3	fishstation	wiredensity	Haul	Haul	Station

nmdbioticv3	fishstation	soaktime	Haul	Haul	Station

nmdbioticv3	fishstation	tripno	Station	Haul	Station

nmdbioticv3	fishstation	fishabundance	Station	Haul	Station

nmdbioticv3	fishstation	fishdistribution	Station	Haul	Station

nmdbioticv3	fishstation	landingsite	Station	Haul	Station

nmdbioticv3	fishstation	fishingground	Station	Haul	Station

nmdbioticv3	fishstation	vesselcount	Station	Haul	Station

nmdbioticv3	fishstation	dataquality	Haul	Haul	Station

nmdbioticv3	fishstation	haulvalidity	Haul	Haul	Station

nmdbioticv3	fishstation	flora	Station	Haul	Station

nmdbioticv3	fishstation	vegetationcover	Station	Haul	Station

nmdbioticv3	fishstation	visibility	Station	Haul	Station

nmdbioticv3	fishstation	waterlevel	Station	Haul	Station

nmdbioticv3	fishstation	winddirection	Station	Haul	Station

nmdbioticv3	fishstation	windspeed	Station	Haul	Station

nmdbioticv3	fishstation	clouds	Station	Haul	Station

nmdbioticv3	fishstation	sea	Station	Haul	Station

nmdbioticv3	fishstation	weather	Station	Haul	Station

nmdbioticv3	fishstation	stationcomment	Station	Haul	Station

nmdbioticv3	catchsample	commonname	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3	catchsample	catchcategory	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3	catchsample	catchpartnumber	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	aphia	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3	catchsample	scientificname	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3	catchsample	identification	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3	catchsample	foreignobject	SpeciesCategory	Sample	SpeciesCategory

nmdbioticv3	catchsample	sampletype	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	group	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	conservation	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	catchproducttype	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	raisingfactor	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	catchweight	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	catchvolume	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	catchcount	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	abundancecategory	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	sampleproducttype	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	lengthmeasurement	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	lengthsampleweight	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	lengthsamplevolume	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	lengthsamplecount	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	specimensamplecount	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	agesamplecount	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	agingstructure	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	parasite	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	stomach	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	intestine	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	tissuesample	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	samplerecipient	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	catchcomment	Sample	Sample	SpeciesCategory

nmdbioticv3	catchsample	catchsampleid	Sample	Sample	SpeciesCategory