Convert landing data

StoX function Convert landing data to the aggregated format StoxLandingData

StoxLanding(
  LandingData,
  appendColumns = character(),
  appendColumnsNames = appendColumns
)

Arguments

LandingData
Sales-notes data. See LandingData

appendColumns
character() vector that identifies additional columns in LandingData to append to StoxLandingData.

appendColumnsNames
character() vector that defines the names of the columns in 'appendColumns' in the output.

Value

StoxLandingData, aggregated landings data.

Details

All columns that are not the ones aggregated (weight), will be used as aggregation variables. This includes any columns added with 'appendColumns' and may not make much sense for continous variables.

If 'LandingData' does not contain columns identified by 'appendColumns'. NA columns will be added.

Correspondances indicate which field a value is derived from, not necessarily verbatim copied.

Correspondance to LandingData (http://www.imr.no/formats/landinger/v2):

Species

Art_kode

Year

Fangstår

CatchDate

SisteFangstdato

Gear

Redskap_kode

Area

Hovedområde_kode

SubArea

Lokasjon_kode

Coastal

KystHav_kode

N62Code

NordSørFor62GraderNord

VesselLengthGroup

Lengdegruppe_kode

CountryVessel

Fartøynasjonalitet_kode

LandingSite

Mottaksstasjon

CountryLanding

Landingsnasjon_kode

Usage

HovedgruppeAnvendelse_kode

RoundWeightKilogram

Rundvekt