StoxLandingData

Table (data.table) with aggregated weight of landings from landing records.

Column definitions

Species

character() code for species category (species identified by market or regulation standards. Several codes may code the same species or stock, and some catch may be recorded only at higher taxonomic classifications)

Year

integer() Year of catch

CatchDate

POSIXct() Date of catch (last catch on trip) in UTC

Gear

character() Code for gear used for catch (dominant gear for trip)

Area

character() Area code for the position where the catch was cought (dominant area for trip)

SubArea

character() Subdivision of area code for the position where the catch was cought (dominant area for trip)

Coastal

character() Code indidcating whether catch was taken within coastal delimitation line (dominant side for trip)

N62Code

character() Code indidcating whether catch was taken north or south of 62 deg. Lat. (dominant side for trip)

VesselLengthGroup

character() Length group for vessel

CountryVessel

character() Country of the vessel that caugth the catch

LandingSite

character() Code identifying landing site (buyer of catch)

CountryLanding

character() Country where catch was landed

Usage

character() Code for market usage of catch.

RoundWeightKilogram

numeric() Weight of round catch in kg.

See also