Stox units

Standard units used in Stox

data(StoxUnits)

Format

A data.table with the following columns:

id

unique identifier for unit

quantity

The quantity the units apply to. Units can be converted to other units of the same quantity

symbol

Symbol or shorthand notation for the unit. Unique within quantity

shortname

Short name of the unit, intended for plotting and menu options

name

Name of the unit

conversion

Conversion factor. All units of the same quantity are expressed relative to common base unit whos value for conversion is 1.

The 'id', 'quantity', and 'conversion' defines a unit. 'symbol', shortname' and 'name' should are unique for a given quantity

See also

Examples

data(StoxUnits)
StoxUnits
#>                                id conversion            quantity    symbol
#>                            <char>      <num>              <char>    <char>
#>  1:                      mass-mcg  1.000e-09                mass       mcg
#>  2:                       mass-mg  1.000e-06                mass        mg
#>  3:                        mass-g  1.000e-03                mass         g
#>  4:                       mass-kg  1.000e+00                mass        kg
#>  5:                      mass-ton  1.000e+03                mass         t
#>  6:                       mass-kt  1.000e+06                mass        kt
#>  7:                       mass-Mt  1.000e+09                mass        Mt
#>  8:                     length-mm  1.000e-03              length        mm
#>  9:                     length-cm  1.000e-02              length        cm
#> 10:                      length-m  1.000e+00              length         m
#> 11:                     length-km  1.000e+03              length        km
#> 12:                    length-nmi  1.852e+03              length         M
#> 13:                 cardinality-N  1.000e+00         cardinality         N
#> 14:                cardinality-kN  1.000e+03         cardinality        kN
#> 15:                cardinality-MN  1.000e+06         cardinality        MN
#> 16:                cardinality-GN  1.000e+09         cardinality        GN
#> 17:                cardinality-TN  1.000e+12         cardinality        TN
#> 18:   area_number_density-N/nmi^2  1.000e+00 area_number_density   N/nmi^2
#> 19:  area_number_density-kN/nmi^2  1.000e+03 area_number_density  kN/nmi^2
#> 20:  area_number_density-MN/nmi^2  1.000e+06 area_number_density  MN/nmi^2
#> 21:  area_number_density-GN/nmi^2  1.000e+09 area_number_density  GN/nmi^2
#> 22:  area_number_density-TN/nmi^2  1.000e+12 area_number_density  GN/nmi^2
#> 23: area_weight_density-mcg/nmi^2  1.000e-09 area_weight_density mcg/nmi^2
#> 24:  area_weight_density-mg/nmi^2  1.000e-06 area_weight_density  mg/nmi^2
#> 25:   area_weight_density-g/nmi^2  1.000e-03 area_weight_density   g/nmi^2
#> 26:  area_weight_density-kg/nmi^2  1.000e+00 area_weight_density  kg/nmi^2
#> 27: area_weight_density-ton/nmi^2  1.000e+03 area_weight_density ton/nmi^2
#> 28:  area_weight_density-kt/nmi^2  1.000e+06 area_weight_density  kt/nmi^2
#> 29:  area_weight_density-Mt/nmi^2  1.000e+09 area_weight_density  Mt/nmi^2
#> 30:                NASC-m^2/nmi^2  1.000e+00                NASC m^2/nmi^2
#> 31:              fraction-decimal  1.000e+00            fraction        0.
#> 32:              fraction-percent  1.000e-02            fraction         %
#> 33:                  frequency-Hz  1.000e+00           frequency        Hz
#> 34:                 frequency-KHz  1.000e+03           frequency       kHz
#> 35:                      age-year  1.000e+00                 age         y
#> 36:                        time-s  1.000e+00                time         s
#> 37:              datetime-ISO8601  1.000e+00            datetime   ISO8601
#> 38:         longitude-degree_east  1.000e+00               angle        °N
#> 39:         latitude-degree_north  1.000e+00               angle        °E
#> 40:                  angle-degree  1.000e+00               angle         °
#>                                id conversion            quantity    symbol
#>                            <char>      <num>              <char>    <char>
#>                   shortname
#>                      <char>
#>  1:                     mcg
#>  2:                      mg
#>  3:                       g
#>  4:                      kg
#>  5:                     ton
#>  6:                 kiloton
#>  7:                 megaton
#>  8:                      mm
#>  9:                      cm
#> 10:                       m
#> 11:                      km
#> 12:                     nmi
#> 13:             individuals
#> 14:        10^3 individuals
#> 15:        10^6 individuals
#> 16:        10^9 individuals
#> 17:       10^12 individuals
#> 18:       individuals/nmi^2
#> 19:  10^3 individuals/nmi^2
#> 20:  10^6 individuals/nmi^2
#> 21:  10^9 individuals/nmi^2
#> 22: 10^12 individuals/nmi^2
#> 23:               mcg/nmi^2
#> 24:                mg/nmi^2
#> 25:                 g/nmi^2
#> 26:                kg/nmi^2
#> 27:               ton/nmi^2
#> 28:                kt/nmi^2
#> 29:                Mt/nmi^2
#> 30:               m^2/nmi^2
#> 31:                fraction
#> 32:                       %
#> 33:                      Hz
#> 34:                     kHz
#> 35:                    year
#> 36:                       s
#> 37:                 ISO8601
#> 38:             degree east
#> 39:            degree north
#> 40:                  degree
#>                   shortname
#>                      <char>
#>                                                                                 name
#>                                                                               <char>
#>  1:                                                                        microgram
#>  2:                                                                        milligram
#>  3:                                                                             gram
#>  4:                                                                         kilogram
#>  5:                                                                       metric ton
#>  6:                                                                   metric kiloton
#>  7:                                                                   metric megaton
#>  8:                                                                       millimeter
#>  9:                                                                       centimeter
#> 10:                                                                            meter
#> 11:                                                                        kilometer
#> 12:                                                                    nautical mile
#> 13:                                                                      individuals
#> 14:                                                             thousand individuals
#> 15:                                                              million individuals
#> 16:                                                              billion individuals
#> 17:                                                             trillion individuals
#> 18:                                             individuals per square nautical mile
#> 19:                                    thousand individuals per square nautical mile
#> 20:                                     million individuals per square nautical mile
#> 21:                                     billion individuals per square nautical mile
#> 22:                                    trillion individuals per square nautical mile
#> 23:                                               microgram per square nautical mile
#> 24:                                               milligram per square nautical mile
#> 25:                                                    gram per square nautical mile
#> 26:                                                kilogram per square nautical mile
#> 27:                                              metric ton per square nautical mile
#> 28:                                          metric kiloton per square nautical mile
#> 29:                                          metric megaton per square nautical mile
#> 30: square meter backscattering cross-section per square nautical mile surveyed area
#> 31:                                                                          decimal
#> 32:                                                                          percent
#> 33:                                                                            hertz
#> 34:                                                                        kilohertz
#> 35:                                                                             year
#> 36:                                                                           second
#> 37:                                                                          ISO8601
#> 38:                                                                WGS84 degree east
#> 39:                                                               WGS84 degree north
#> 40:                                                                           degree
#>                                                                                 name
#>                                                                               <char>