Translate AcousticData

This function translates one or more columns of AcousticData to new values given by the input Translation.

TranslateAcoustic(
  AcousticData,
  TranslationDefinition = c("FunctionParameter", "FunctionInput"),
  TranslationTable = data.table::data.table(),
  VariableName = character(),
  Conditional = FALSE,
  ConditionalVariableNames = character(),
  Translation,
  PreserveClass = TRUE
)

Arguments

AcousticData

AcousticData.

TranslationDefinition

The method to use for defining the Translation, one of FunctionParameter to define the Translation on the fly in this function using the TranslationTable, or FunctionInput to import Translation process data from a previously run process using the function DefineTranslation. When reading the translation from a file FunctionInput must be used.

TranslationTable

A table of the variable to translate in the first column; the column NewValue giving the values to translate to in the second column; followed by zero or more conditional variables. Use NA in the Value column to translate missing values (shown as "-" in View output in the StoX GUI, and usually as empty cell in excel). Values in the TranslationTable can be given either as single values or as expressions or functions of the variable specified by the column name. See details.

VariableName

An optional string naming the variable to translate, which can be given if only one variable should be translated. If more than one variable should be translated using the same Translation, a column named "VariableName" must be inclcuded in TranslationTable (possibly read from the FileName) AND VariableName must be an empty string (blank field in the GUI).

Conditional

Logical: If TRUE the columns ConditionalVariableNames and ConditionalValue are expected in the TranslationTable. These define a variable interacting with the VariableName and Value, so that VariableName is changed from Value to NewValue only when ConditionalVariableNames has the value given by ConditionalValue. Note that ConditionalVariableNames must exist in the same table as VariableName.

ConditionalVariableNames

Similar to VariableName, but naming the conditional variables.

Translation

Translation.

PreserveClass

Logical: If TRUE (the default) do not convert the class of the data by the class of the translation table. E.g., with the default (PreserveClass = TRUE), if the variable to translate is integer (1, 2, etc) and the NewValue is character ("First", "Second", etc), the NewValue will be converted to integer before translation, resulting in NA if the character strings are not convertible to integer. In this example it could be the user's intention to convert the class of the variable to translate instead, which is possible using PreserveClass = FALSE.

Value

A AcousticData object.