Convert BioticData to ICESBiotic format

Given an BioticData object, this function converts to ICESBiotic format. Note that this function only supports BioticData generated from NMDBiotic version > 3 XML files.

ICESBiotic(
  BioticData,
  SurveyName = character(),
  Country = character(),
  Organisation = integer(),
  AllowRemoveSpecies = TRUE
)

Arguments

BioticData

a BioticData object from an XML file with NMD biotic version 3 format.

SurveyName

A string naming the survey. Must be one of the names listed on https://vocab.ices.dk/?ref=1453 or NONE.

Country

The ISO_3166 code of the country running the cruise. See http://vocab.ices.dk/?ref=337.

Organisation

An integer code representing the organization running the cruise. See https://vocab.ices.dk/?ref=1398 for a list of possible codes (e.g., Institute of Marine Research: 612).

AllowRemoveSpecies

ICES submission will not allow the resulting CSV file to be uploaded if the file contains species not listed in https://acoustic.ices.dk/Services/Schema/XML/SpecWoRMS.xml. Setting this parameter to TRUE will remove the unlisted species records.

Value

An ICESBioticData object.