StoxBiotic data format.

The StoxBiotic data format is defined by StoX as a common format to which data from different biotic sampling formats are converted, guaranteeing consistent interpretation and documentation of all its variables.

Details

The StoxBiotic format is defined according to the general sampling hierarchy of StoX which is used as a basis for both the StoxcBiotic and StoxAcoustic format. The variables of the StoxBiotic format are given by the tables below, where the columns NMDBiotic and ICESBiotic show the origin of variable in BioticData read from NMDBiotic https://www.imr.no/formats/nmdbiotic/ and ICESBiotic https://www.ices.dk/data/data-portals/Pages/acoustic.aspx files, respectively (see BioticData). Variables with non-unique names across tables of the BioticData are written as variableName.tableName, and "-" and "/" denotes concatenation for character type variables with "-" and "/" as separation character, respectively. Units are given with shortname of the StoxUnits:

Cruise level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

CruiseKey	Key of the Cruise table	None	Character	"2021105"	cruise/missiontype/startyear/platform/missionnumber	LocalID

Cruise	Unique Cruise identifier	None	Character	"2021105"	Same as CruiseKey. Can be translated.	Same as CruiseKey. Can be translated.

Platform	Data collection platform identifier	None	Character	"1019"	platform	Platform

Station level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

StationKey	Key of the Station level	None	Character	"1"	station	StationName

Station	Unique Station identifier	None	Character	"2021105-1"	CruiseKey-StationKey	CruiseKey-StationKey

CatchPlatform	Platform performing the actual sampling (can be different from the data collection platform)	None	Character	"1019"	catchplatform (platform.fishstation for NMDBiotic 1.1 and 1.4)	Platform

DateTime	UTC time at start of the station, stored as POSIXct	ISO8601	Character	2020-09-09T01:02:03.456Z	stationstartdate + stationstarttime (startdate.fishstation + starttime.fishstation for NMDBiotic 1.1 and 1.4)	StartTime

Longitude	Longitude at start of the station	degree east	Numeric	62.5	longitudestart	StartLongitude

Latitude	Latitude at start of the station	degree north	Numeric	5.1	latitudestart	StartLatitude

BottomDepth	BottomDepth at the station, given directly in ICESBitoic files and calculated as the average of bottom depth at start and end of the station in NMDBiotic files	m	Numeric	123.2	(bottomdepthstart + bottomdepthstop) / 2	BottomDepth

Haul level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

HaulKey	Key of the Haul level	None	Character	"2"	serialnumber (serialno for NMDBiotic 1.1 and 1.4)	Number

Haul	Unique Haul identifier	None	Character	"2021105-1-2"	CruiseKey-StationKey-HaulKey	CruiseKey-StationKey-HaulKey

Gear	Identifier of the gear	None	Character	"3270"	gear	Gear

TowDistance	Distance between start and end of the haul	nmi	Numeric	1.5	distance	Distance / 1852

EffectiveTowDistance	Effective tow distance	nmi	Numeric	1.5	Same as TowDistance. Can be translated.	Same as TowDistance. Can be translated.

MinHaulDepth	Minimum depth of the haul (trawl headline)	m	Numeric	65	fishingdepthmin	MinTrawlDepth

MaxHaulDepth	Maximum depth of the haul (trawl headline)	m	Numeric	35	fishingdepthmax	MaxTrawlDepth

VerticalNetOpening	Vertical span of the net	m	Numeric	23	verticaltrawlopening (trawlopening for NMDBiotic 1.1 and 1.4)	Netopening

HorizontalNetOpening	Vertical span of the net	m	Numeric	105	wingspread (missing (NA) for NMDBiotic 1.1 and 1.4)	WingSpread

TrawlDoorSpread	Distance between the trawl doors	m	Numeric	125	trawldoorspread (doorspread for NMDBiotic 1.1 and 1.4)	DoorSpread

SpeciesCategory level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

SpeciesCategoryKey	Key of the SpeciesCategory level	None	Character	"sild'G03/161722.G03/126417/Clupea harengus"	commonname/catchcategory/aphia/scientificname (noname/species/aphia/group for NMDBiotic 1.1 and 1.4)	SpeciesCode

SpeciesCategory	The species category	None	Character	"Herring"	Same as SpeciesCategoryKey. Can be translated.	Same as SpeciesCategoryKey. Can be translated.

Sample level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

SampleKey	Key of the Sample level	None	Character	"1"	catchsampleid (samplenumber for NMDBiotic 1.1 and 1.4)	SpeciesCategory

Sample	Unique Sample identifier	None	Character	"2021105-1-2-sild'G03/161722.G03/126417/Clupea harengus-1"	CruiseKey-StationKey-HaulKey-SpeciesCategoryKey-SampleKey	CruiseKey-StationKey-HaulKey-SpeciesCategoryKey-SampleKey

CatchFractionWeight	Total weight of the catch for the SpeciesCategory and sub category (e.g., fractions such as juveniles and adults)	kg	Numeric	49.9	catchweight (missing (NA) if catchproducttype != 1) (weight (missing (NA) if producttype != 1) for NMDBiotic 1.1 and 1.4)	SpeciesCategoryWeight (taking WeightUnit into account)

CatchFractionNumber	Total number of individuals of the catch for the SpeciesCategory and sub category (e.g., fractions such as juveniles and adults)	individuals	Integer	295	catchcount (count for NMDBiotic 1.1 and 1.4)	SpeciesCategoryNumber

SampleWeight	Total weight of the sample for individual measurements	kg	Numeric	4.6	lengthsampleweight (missing (NA) if sampleproducttype != 1)	SubsampleWeight (taking WeightUnit into account)

SampleNumber	Size of the sample for individual measurements	individuals	Integer	100	lengthsamplecount	SubsampledNumber

Individual level:

Variable	Description	Unit	Data type	Example	NMDBiotic	ICESBiotic

IndividualKey	Key of the Individual level	None	Character	"2"	specimenid (specimenno for NMDBiotic 1.1 and 1.4)	FishID

Individual	Unique Individual identifier	None	Character	"2021105-1-2-sild'G03/161722.G03/126417/Clupea harengus-1-2"	CruiseKey-StationKey-HaulKey-SpeciesCategoryKey-SampleKey-IndividualKey	CruiseKey-StationKey-HaulKey-SpeciesCategoryKey-SampleKey-IndividualKey

IndividualRoundWeight	Round weight (the whole fish) fo the individual	g	Numeric	123	individualweight * 1000 (missing (NA) if individualproducttype != 1) (weight.individual (missing (NA) if producttype.individual != 1) for NMDBiotic 1.1 and 1.4)	IndividualWeight (taking WeightUnit.Biology into account)

IndividualTotalLength	Total length (from snout to end of fin), given as the lower end of the interval of width given by LengthResolution	cm	Numeric	14.5	length * 100 (missing (NA) if lengthmeasurement != "E")	LengthClass.Biology (taking LengthCode.Biology into account)

LengthResolution	Resolution of IndividualTotalLength	cm	Numeric	0.5	lengthresolution (lengthunit for NMDBiotic 1.1 and 1.4) converted to cm	LengthCode.Biology converted to cm

WeightMeasurement	Specification of how IndividualRoundWeight was measured; one of "IndividualWeight" and "AverageWeight". For data from NMDBiotic files this is always "IndividualWeight".	None	Character	"IndividualWeight"	"IndividualWeight"	IndividualWeight if IndividualWeight is given, "AverageWeight" if WeightAtLength and NumberAtLength are given

IndividualAge	Age of an individual	year	Numeric	3	age (located in the agedetermination level) for the preferredagereading	IndividualAge

IndividualSex	sex of an individual	F is female, M is male	Character	"F"	"F" if sex = 1, "M" if sex = 2	IndividualSex