Informative match.arg

Informative match.arg

match_arg_informative(
  arg,
  choices,
  several.ok = FALSE,
  arg_name = substitute(arg),
  ignore.case = FALSE
)

Arguments

arg

a character vector (of length one unless several.ok is TRUE) or NULL which means to take choices[1].

choices

a character vector of candidate values, often missing, see ‘Details’.

several.ok

logical specifying if arg should be allowed to have more than one element.

arg_name

The name of the argument, interpreted by default.

ignore.case

Logical: Should case be igngored.